Follow Precede 2, still need test on main grid
I made a mistake in preceding columns/rows since last commit. I meant to subtract 1 from row number on vertical and from column on vertical. So I fixed that error.
But Graham and I still need to call the shutoFFHorizontalFollow, Vertical, Precede on the grid to get the right letter from the crossword puzzle